A2/requirements.txt
altgraph==0.17.2
anyio==3.5.0
asgiref==3.5.0
blinker==1.6.2
boto3==1.26.141
botocore==1.29.141
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
distlib==0.3.4
fastapi==0.75.0
filelock==3.6.0
Flask==2.3.2
grpcio==1.36.1
grpcio-tools==1.54.2
h11==0.13.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
macholib==1.15.2
MarkupSafe==2.1.2
numpy==1.22.3
packaging==21.3
pipenv==2022.1.8
platformdirs==2.5.1
protobuf==4.23.2
pydantic==1.9.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
s3transfer==0.6.1
sacremoses==0.0.49
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tokenizers==0.11.6
torch==1.11.0
tqdm==4.63.0
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.13.3
virtualenv-clone==0.5.7
Werkzeug==2.3.4
zipp==3.15.0
